requirements.txt
Flask==3.1.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.3
anthropic==0.40.0
openai==1.58.1
python-dotenv==1.0.1
bleach==6.2.0
playwright==1.49.1
